chore: migrated Header and Title component props into HeaderComponentModels.ts
